cloudmonkey: have filter applicable for list outputs as well
If filter is used for list output (default output) now, it filters out keys
based on the filter passed. If only one filter is applied say 'id', then only
the values are printed. For example;

  $ cloudmonkey list domains filter=id
  111b3bf9-4234-11e4-be95-00012e4ff7d0

In the above example, only UUIDs are printed. This can be useful for people to
just grab uuids and do interesting things with it.

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>