Don't force connections to TLSv1 (#25)
Forcing the connection to TLSv1 is actually downgrading connections that would otherwise be TLSv1_2. Cloudmonkey is better off auto-negotiating ssl/tls versions and versions should be enforced on the server side. Since the Go version is still alpha, it would be very useful to have the Python version play nice with TLSv1_2.